requirements.txt
Flask==3.1.2
requests==2.32.0
beautifulsoup4==4.13.3
python-dotenv==1.1.0